02_FastAPI_Interface/requirements.txt
# FastAPI dan dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Streamlit
streamlit==1.28.1

# Computer Vision dan Machine Learning
opencv-python==4.8.1.78
dlib==19.24.2
imutils==0.5.4
scipy==1.11.4
numpy==1.24.3

# Image processing
Pillow==10.1.0

# Audio untuk alarm (opsional)
pygame==2.5.2

# HTTP requests untuk Streamlit
requests==2.31.0

# Data visualization
plotly==5.17.0
pandas==2.1.3

# Logging
logging